+++ server/src/main/java/org/opensearch/search/dfs/DfsPhase.java
@@ -35,10 +35,10 @@ package org.opensearch.search.dfs;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.search.CollectionStatistics;
 import org.apache.lucene.search.IndexSearcher;
-import org.apache.lucene.search.Query;
 import org.apache.lucene.search.ScoreMode;
 import org.apache.lucene.search.TermStatistics;
 import org.opensearch.core.tasks.TaskCancelledException;
+import org.opensearch.index.query.ParsedQuery;
 import org.opensearch.search.internal.SearchContext;
 import org.opensearch.search.rescore.RescoreContext;
 
@@ -86,7 +86,7 @@ public class DfsPhase {
 
             searcher.createWeight(context.searcher().rewrite(context.query()), ScoreMode.COMPLETE, 1);
             for (RescoreContext rescoreContext : context.rescore()) {
-                for (org.opensearch.index.query.ParsedQuery parsedQuery : rescoreContext.getParsedQueries()) {
+                for (ParsedQuery parsedQuery : rescoreContext.getParsedQueries()) {
                     searcher.createWeight(context.searcher().rewrite(parsedQuery.query()), ScoreMode.COMPLETE, 1);
                 }
             }